gallery section CSS fundamentals
Relates #9 I've created the gir layout and populated with a few images. Will still edit a bit more but this commit also includes a few adjustments to the page overall - I added h2 elements to each section for better navigation and refined the look slightly. I'll add some info on each picture in the gallery to be shown when we hover next and tomorrow should start testing everything for accessibility