screen reader and mouseless accessibility improved
Relates #20 and Relates#19 I have added aria attributes to allow for a clear navigation with the screen reader and tabIndex to a few elements for mouseless navigation. Lighthouse is currently showing 89% but it says a few aria-attributes are incorrect, however, while testing I see that the page is much less accessible without them. It's probably worth you running a report to see what this is about. I also adjusted some colours for better contrast and the carrousell should now be a bit clearer. It's probably worth mergin this even if we don't close the issue right now (unless you want to suggest changes) so we can pull the main with the current form and the accessibility changes - that way we can keep working on accessibility with the most updated look of the page.